refactor: separate runner orchestration concerns
