+++ crates/bazel-mcp-reducer/src/lib.rs
@@ -7,6 +7,7 @@ mod extension;
 mod query;
 mod starlark;
 mod test;
+mod test_evidence;
 mod text;
 
 pub use budget::{Budget, Budgeted};
@@ -25,4 +26,5 @@ pub use starlark::{
     REDUCER_API_VERSION, StarlarkLimits, StarlarkReducerConfig, load_starlark_reducers,
 };
 pub use test::{TestFailureAccumulator, TestFailureEvidence, TestXmlError, parse_test_xml};
+pub use test_evidence::{TestEvidenceInput, TestEvidenceReducer, TestEvidenceResult};
 pub use text::{deduplicate_lines, normalize_terminal_text};
