+++ crates/bazel-mcp-runner/src/lib.rs
@@ -1,13 +1,21 @@
 //! Asynchronous Bazel invocation lifecycle and application service.
 
+mod artifacts;
 mod cancel;
 mod capture;
+mod evidence;
+mod execution;
+mod inspection;
 mod output_base_lock;
+mod scheduler;
 mod service;
+mod test_evidence;
 mod version;
 
 pub use bazel_mcp_reducer::{StarlarkLimits, StarlarkReducerConfig};
+pub use inspection::{InspectRequest, InspectResult, InspectView};
+#[doc(hidden)]
+pub use service::RunnerTestSupport;
 pub use service::{
-    BepTransport, CancelResult, InspectRequest, InspectResult, InspectView, InvocationProgress,
-    InvocationService, RunnerConfig, RunnerError,
+    BepTransport, CancelResult, InvocationProgress, InvocationService, RunnerConfig, RunnerError,
 };
